Move font loading to the FontLoader
