+++ apps/openmw/mwgui/windowmanagerimp.cpp
@@ -181,7 +181,6 @@ namespace MWGui
       , mRestAllowed(true)
       , mShowOwned(0)
       , mEncoding(encoding)
-      , mFontHeight(16)
       , mVersionDescription(versionDescription)
       , mWindowVisible(true)
     {
@@ -218,13 +217,6 @@ namespace MWGui
         ItemWidget::registerComponents();
         SpellView::registerComponents();
         Gui::registerAllWidgets();
-
-        int fontSize = Settings::Manager::getInt("font size", "GUI");
-        fontSize = std::min(std::max(12, fontSize), 20);
-        mFontHeight = fontSize;
-
-        MyGUI::ResourceManager::getInstance().unregisterLoadXmlDelegate("Resource");
-        MyGUI::ResourceManager::getInstance().registerLoadXmlDelegate("Resource") = newDelegate(this, &WindowManager::loadFontDelegate);
 
         MyGUI::FactoryManager::getInstance().registerFactory<MWGui::Controllers::ControllerFollowMouse>("Controller");
 
@@ -280,94 +272,6 @@ namespace MWGui
         mVideoWrapper = new SDLUtil::VideoWrapper(window, viewer);
         mVideoWrapper->setGammaContrast(Settings::Manager::getFloat("gamma", "Video"),
                                         Settings::Manager::getFloat("contrast", "Video"));
-    }
-
-    void WindowManager::loadFontDelegate(MyGUI::xml::ElementPtr _node, const std::string& _file, MyGUI::Version _version)
-    {
-        MyGUI::xml::ElementEnumerator resourceNode = _node->getElementEnumerator();
-        bool createCopy = false;
-        while (resourceNode.next("Resource"))
-        {
-            std::string type, name;
-            resourceNode->findAttribute("type", type);
-            resourceNode->findAttribute("name", name);
-
-            if (name.empty())
-                continue;
-
-            if (Misc::StringUtils::ciEqual(type, "ResourceTrueTypeFont"))
-            {
-                createCopy = true;
-
-                // For TrueType fonts we should override Size and Resolution properties
-                // to allow to configure font size via config file, without need to edit XML files.
-                // Also we should take UI scaling factor in account.
-                int resolution = Settings::Manager::getInt("ttf resolution", "GUI");
-                resolution = std::min(960, std::max(48, resolution));
-
-                float uiScale = Settings::Manager::getFloat("scaling factor", "GUI");
-                resolution *= uiScale;
-
-                MyGUI::xml::ElementPtr resolutionNode = resourceNode->createChild("Property");
-                resolutionNode->addAttribute("key", "Resolution");
-                resolutionNode->addAttribute("value", std::to_string(resolution));
-
-                MyGUI::xml::ElementPtr sizeNode = resourceNode->createChild("Property");
-                sizeNode->addAttribute("key", "Size");
-                sizeNode->addAttribute("value", std::to_string(mFontHeight));
-            }
-            else if (Misc::StringUtils::ciEqual(type, "ResourceSkin") ||
-                     Misc::StringUtils::ciEqual(type, "AutoSizedResourceSkin"))
-            {
-                // We should adjust line height for MyGUI widgets depending on font size
-                MyGUI::xml::ElementPtr heightNode = resourceNode->createChild("Property");
-                heightNode->addAttribute("key", "HeightLine");
-                heightNode->addAttribute("value", std::to_string(mFontHeight+2));
-            }
-        }
-
-        MyGUI::ResourceManager::getInstance().loadFromXmlNode(_node, _file, _version);
-
-        if (createCopy)
-        {
-            MyGUI::xml::ElementPtr copy = _node->createCopy();
-
-            MyGUI::xml::ElementEnumerator copyFont = copy->getElementEnumerator();
-            while (copyFont.next("Resource"))
-            {
-                std::string type, name;
-                copyFont->findAttribute("type", type);
-                copyFont->findAttribute("name", name);
-
-                if (name.empty())
-                    continue;
-
-                if (Misc::StringUtils::ciEqual(type, "ResourceTrueTypeFont"))
-                {
-                    // Since the journal and books use the custom scaling factor depending on resolution,
-                    // setup separate fonts with different Resolution to fit these windows.
-                    // These fonts have an internal prefix.
-                    int resolution = Settings::Manager::getInt("ttf resolution", "GUI");
-                    resolution = std::min(960, std::max(48, resolution));
-
-                    float currentX = Settings::Manager::getInt("resolution x", "Video");
-                    float currentY = Settings::Manager::getInt("resolution y", "Video");
-                    // TODO: read size from openmw_layout.xml
-                    float heightScale = (currentY / 520);
-                    float widthScale = (currentX / 600);
-                    float uiScale = std::min(widthScale, heightScale);
-                    resolution *= uiScale;
-
-                    MyGUI::xml::ElementPtr resolutionNode = copyFont->createChild("Property");
-                    resolutionNode->addAttribute("key", "Resolution");
-                    resolutionNode->addAttribute("value", std::to_string(resolution));
-
-                    copyFont->setAttribute("name", "Journalbook " + name);
-                }
-            }
-
-            MyGUI::ResourceManager::getInstance().loadFromXmlNode(copy, _file, _version);
-        }
     }
 
     void WindowManager::loadUserFonts()
@@ -566,7 +470,7 @@ namespace MWGui
 
     int WindowManager::getFontHeight() const
     {
-        return mFontHeight;
+        return mFontLoader->getFontHeight();
     }
 
     void WindowManager::setNewGame(bool newgame)
@@ -587,7 +491,6 @@ namespace MWGui
         {
             mKeyboardNavigation.reset();
 
-            MyGUI::ResourceManager::getInstance().unregisterLoadXmlDelegate("Resource");
             MyGUI::LanguageManager::getInstance().eventRequestTag.clear();
             MyGUI::PointerManager::getInstance().eventChangeMousePointer.clear();
             MyGUI::InputManager::getInstance().eventChangeKeyFocus.clear();
